+++ mobile/android/android-components/docs/changelog.md
@@ -10,6 +10,9 @@ permalink: /changelog/
   * Added `StartForegroundService` to safely start a foreground service, see [Bug 1839039](https://bugzilla.mozilla.org/show_bug.cgi?id=1839039) for crash reference.
   * Added `ProcessInfoProvider` and `BuildVersionProvider` to get information about the app process and the build version.
 
+* **browser-engine-gecko**
+    * For screenshot capture, include exception in failure result rather than throwing.
+
 # 126.0
 
 * **browser-menu**